requirements.txt
Flask==3.0.3
azure-ai-vision==1.0.0b3
azure-core==1.30.2
requests==2.32.3
gunicorn
